pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imputify"
version = "0.1.0"
description = "A library for imputation of missing data in tabular datasets with comprehensive evaluation metrics"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Gabriel Francisco dos Santos Silva", email = "gabfssilva@gmail.com" }
]
keywords = [
    "missing-data",
    "imputation",
    "machine-learning",
    "deep-learning",
    "autoencoder",
    "scikit-learn",
    "tabular-data",
    "data-preprocessing",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=2.0",
    "pandas>=2.0",
    "pyampute>=0.0.3",
    "scikit-learn>=1.5",
    "torch>=2.0",
    "accelerate>=0.25",
    "peft>=0.7",
    "transformers>=4.36",
    "jsonpickle>=3.0",
    "xgboost>=3.1.3",
]

[project.optional-dependencies]
cpu = ["torch>=2.0"]
gpu = ["torch>=2.0"]
quantization = ["optimum-quanto>=0.2"]
cuda = ["bitsandbytes>=0.41", "hf_transfer>=0.1"]
experiments = [
    "skypilot[runpod]>=0.11.1",
    "cyclopts>=4.5.1",
    "rich>=13.0",
    "modal>=1.0",
    "plotext>=5.0",
]
types = ["pandas-stubs>=2.0"]
all = ["imputify[quantization,cuda,experiments,types]"]

[tool.hatch.build.targets.wheel]
packages = ["imputify"]

[project.urls]
Homepage = "https://github.com/gabfssilva/imputify"
Repository = "https://github.com/gabfssilva/imputify"
Issues = "https://github.com/gabfssilva/imputify/issues"

[tool.uv]
conflicts = [
    [{ extra = "cpu" }, { extra = "gpu" }],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-gpu", extra = "gpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
